# server/openapi.yaml
openapi: 3.1.0
info:
  title: Harness Server API
  version: 0.1.0
  description: |
    The HTTP surface `harness serve` exposes inside a sandbox. It is a
    protocol, not a product: no static assets, no UI — a single external
    orchestrator drives many harness instances through this API.

    ## Design notes

    - **The session log is the wire contract.** Durable records (messages,
      model swaps, session lifecycle) carry monotonic sequence numbers and
      are replayable; live deltas (text, reasoning, tool progress) stream
      between them and are not persisted. A client that reconnects with
      `from=<last seq>` reconstructs exactly what it missed.
    - **Session status is engine truth.** `busy` means a prompt is being
      processed by the engine — it has nothing to do with whether any client
      is connected. (This deliberately fixes the opencode `/session/status`
      footgun where idle-but-usable sessions reported as absent.)
    - **Sessions outlive the process.** Session logs persist on disk; a
      prompt to a session that is not in memory transparently resumes it
      from its log. "Stored session gone after sandbox restart" is not an
      error mode — orchestrators keep one session ID per task and reuse it.
    - **Prompting is always asynchronous.** The endpoint keeps opencode's
      `prompt_async` name so existing orchestration code ports with a URL
      and auth swap; there is no synchronous variant.
...
servers:
  - url: http://localhost:4096
...
security:
  - runToken: []
components:
  securitySchemes:
    runToken:
      type: http
      scheme: bearer
      description: |
        The workspace's run token — the same per-sandbox token gatekeeper
        scopes credentials with. Compared in constant time. The box proxy
        terminates human/service auth in front of this.
  schemas:
...
    Session:
      type: object
      required: [id, created_at, model, status, messages, workdir]
      properties:
        id:
          type: string
          example: ses_77aa9a7c6f323a7b
        created_at:
          type: string
          format: date-time
        model:
          $ref: "#/components/schemas/ModelRef"
        status:
          $ref: "#/components/schemas/SessionStatus"
        messages:
          type: integer
          description: Count of messages in the session log.
        seq:
          type: integer
          description: Sequence number of the latest durable record.
        goal:
          $ref: "#/components/schemas/GoalSummary"
          description: >
            The session's goal summary, present only when a goal has been set
            for this session in the current server process (in-memory, like the
            /request snapshot). Absent on a freshly reloaded session until a
            goal runs.
        workdir:
          type: string
          description: >
            The session's working directory, durable across resume (persisted
            in the session log header). Defaults to the serve process's
            current working directory when POST /session omitted workdir.
          example: /workspace/task-42
    GoalSummary:
      type: object
      description: Per-session goal state surfaced on Session.goal.
      required: [condition, active, turns]
      properties:
        condition:
          type: string
          description: The goal's completion condition.
        active:
          type: boolean
          description: >
            True while the goal is set but neither achieved nor cleared.
        turns:
          type: integer
          description: Worker turns taken so far.
        last_reason:
          type: string
          description: The most recent evaluator reason (eval or achieved).
    SessionStatus:
      type: string
      enum: [idle, busy]
      description: >
        Engine truth: busy while a prompt is being processed, idle
        otherwise. Independent of client connections.
    ModelRef:
      type: string
      description: Compact "provider/model" form, or a config alias.
      example: anthropic/claude-fable-5
...
    GoalRequest:
      type: object
      required: [condition]
      properties:
        condition:
          type: string
          description: The natural-language completion condition to pursue.
        max_turns:
          type: integer
          description: Maximum worker turns; 0 or omitted means unlimited.
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
  parameters:
    sessionID:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: Missing or invalid run token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No session log with that ID.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
paths:
...
  /session:
    get:
      operationId: listSessions
      summary: List sessions (in-memory and persisted).
      responses:
        "200":
          description: All sessions known to this instance's session dir.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: createSession
      summary: Create a session.
      description: >
        workdir is optional: when given it must clean-resolve (absolute,
        cleaned — ".." cannot escape by string trickery) to a path equal to,
        or nested under, one of the server's configured workspace roots;
        otherwise the request is rejected with 400. Omitted, it defaults to
        the serve process's own current working directory. workdir is durable
        — it round-trips through the session log and survives resume.
        share_workdir opts this session out of the workdir-busy exclusivity
        rule (see prompt_async's 409): by default, two running sessions may
        never share a workdir at the same time.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                model: { $ref: "#/components/schemas/ModelRef" }
                workdir:
                  type: string
                  description: >
                    Optional explicit working directory; see the operation
                    description for validation and defaulting rules.
                share_workdir:
                  type: boolean
                  description: >
                    Opt out of the workdir-busy exclusivity rule for this
                    session (see prompt_async's 409 response).
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400":
          description: >
            workdir does not clean-resolve under an allowed workspace root.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /session/{id}/goal:
    post:
      operationId: startGoal
      summary: Start a goal loop on a session.
      description: |
        Prompts the condition, then after every turn asks an independent,
        tool-less evaluator model (config goal_evaluator_model) whether it is
        met, feeding the evaluator's reason back as guidance until met or the
        turn budget runs out. Runs in the background like prompt_async; progress
        and the durable goal.* records arrive on /event. The goal occupies the
        session's single run slot, so a concurrent prompt_async or goal is 409.
        400 when goal_evaluator_model is not configured, or the condition is
        empty.
      parameters:
        - $ref: "#/components/parameters/sessionID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GoalRequest" }
      responses:
        "202":
          description: Accepted. Body carries the seq to tail events from.
          content:
            application/json:
              schema:
                type: object
                required: [seq]
                properties:
                  seq: { type: integer }
        "400":
          description: No evaluator configured, or empty condition.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: >
            Session is busy with another prompt or goal, OR another
            currently-running session holds the same workdir and neither
            session set share_workdir on creation (the error names the
            holder session).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      operationId: clearGoal
      summary: Cancel the active goal loop, if any.
      description: >
        Cancels the goal loop's context (stopping further turns) and journals a
        durable goal.cleared. Idempotent: a known session with no active goal is
        still 204 (and journals nothing).
      parameters:
        - $ref: "#/components/parameters/sessionID"
      responses:
        "204": { description: Cleared (or nothing was active). }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...